+++ src/Aws/DataPipeline/Resources/datapipeline-2012-10-29.php
@@ -100,6 +100,15 @@ return array (
                         ),
                     ),
                 ),
+                'startTimestamp' => array(
+                    'type' => array(
+                        'object',
+                        'string',
+                        'integer',
+                    ),
+                    'format' => 'date-time',
+                    'location' => 'json',
+                ),
             ),
             'errorResponses' => array(
                 array(
@@ -255,6 +264,59 @@ return array (
                 ),
             ),
             'errorResponses' => array(
+                array(
+                    'reason' => 'An internal service error occurred.',
+                    'class' => 'InternalServiceErrorException',
+                ),
+                array(
+                    'reason' => 'The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven\'t exceeded any of the service limits for your account.',
+                    'class' => 'InvalidRequestException',
+                ),
+            ),
+        ),
+        'DeactivatePipeline' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'DataPipeline.DeactivatePipeline',
+                ),
+                'pipelineId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                    'minLength' => 1,
+                    'maxLength' => 1024,
+                ),
+                'cancelActive' => array(
+                    'type' => 'boolean',
+                    'format' => 'boolean-string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'The specified pipeline was not found. Verify that you used the correct user and account identifiers.',
+                    'class' => 'PipelineNotFoundException',
+                ),
+                array(
+                    'reason' => 'The specified pipeline has been deleted.',
+                    'class' => 'PipelineDeletedException',
+                ),
                 array(
                     'reason' => 'An internal service error occurred.',
                     'class' => 'InternalServiceErrorException',