+++ src/Aws/ElasticBeanstalk/Resources/elasticbeanstalk-2010-12-01.php
@@ -66,6 +66,41 @@ return array (
         ),
     ),
     'operations' => array(
+        'AbortEnvironmentUpdate' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\QueryCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Action' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => 'AbortEnvironmentUpdate',
+                ),
+                'Version' => array(
+                    'static' => true,
+                    'location' => 'aws.query',
+                    'default' => '2010-12-01',
+                ),
+                'EnvironmentId' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                ),
+                'EnvironmentName' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'minLength' => 4,
+                    'maxLength' => 23,
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Unable to perform the specified operation because the user does not have enough privileges for one of more downstream aws services',
+                    'class' => 'InsufficientPrivilegesException',
+                ),
+            ),
+        ),
         'CheckDNSAvailability' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -1424,6 +1459,11 @@ return array (
                     'minLength' => 1,
                     'maxLength' => 100,
                 ),
+                'SolutionStackName' => array(
+                    'type' => 'string',
+                    'location' => 'aws.query',
+                    'maxLength' => 100,
+                ),
                 'OptionSettings' => array(
                     'type' => 'array',
                     'location' => 'aws.query',
@@ -1536,6 +1576,10 @@ return array (
         ),
     ),
     'models' => array(
+        'EmptyOutput' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+        ),
         'CheckDNSAvailabilityResultMessage' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -1738,6 +1782,10 @@ return array (
                     'type' => 'string',
                     'location' => 'xml',
                 ),
+                'AbortableOperationInProgress' => array(
+                    'type' => 'boolean',
+                    'location' => 'xml',
+                ),
                 'Health' => array(
                     'type' => 'string',
                     'location' => 'xml',
@@ -1801,10 +1849,6 @@ return array (
                     'location' => 'xml',
                 ),
             ),
-        ),
-        'EmptyOutput' => array(
-            'type' => 'object',
-            'additionalProperties' => true,
         ),
         'ApplicationVersionDescriptionsMessage' => array(
             'type' => 'object',
@@ -2164,6 +2208,9 @@ return array (
                             ),
                             'Status' => array(
                                 'type' => 'string',
+                            ),
+                            'AbortableOperationInProgress' => array(
+                                'type' => 'boolean',
                             ),
                             'Health' => array(
                                 'type' => 'string',